requirements.txt
click==7.1.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
python-dotenv==0.12.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
WTForms==2.2.1
